+++ library/src/main/java/ibt/ortc/api/Authentication.java
@@ -4,6 +4,8 @@
 package ibt.ortc.api;
 
 //import java.io.BufferedWriter;
+import ibt.ortc.util.IOUtil;
+
 import java.io.IOException;
 //import java.io.OutputStream;
 import java.io.OutputStreamWriter;
