+++ library/src/main/java/ibt/ortc/api/RestWebservice.java
@@ -1,5 +1,7 @@
 package ibt.ortc.api;
 
+import ibt.ortc.util.IOUtil;
+
 import java.io.*;
 import java.net.URL;
